work: Move EventSource to production code
