+++ dom/event/event_source_test.go
@@ -11,13 +11,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-const DEFAULT_BUF = 16
-
 type EventChan chan *event.Event
-
-type EventStream struct {
-	EventChan
-}
 
 func NewEventStream(tgt event.EventTarget, t string, ctx context.Context) EventChan {
 	c := make(chan *event.Event)
@@ -40,9 +34,9 @@ func TestEventsAreReceivedInOrder(t *testing.T) {
 		ctx, cancel := context.WithCancel(t.Context())
 		defer cancel()
 
-		src := EventSource{event.NewEventTarget()}
+		src := event.EventSource{event.NewEventTarget()}
 		const buf = 32
-		c := src.Listen(ctx, "gost-event", BufSize(buf))
+		c := src.Listen(ctx, "gost-event", event.BufSize(buf))
 
 		for i := range buf << 1 {
 			// Dispatch twice as many events as the buffer size
@@ -77,7 +71,7 @@ func TestEventStreamSource(t *testing.T) {
 			newCtx, cancel := context.WithCancel(t.Context())
 
 			spy := &EventTargetSpy{EventTarget: target}
-			src := EventSource{spy}
+			src := event.EventSource{spy}
 			events := src.Listen(newCtx, "gost-event")
 
 			assert.Equal(t, 1, spy.addCallCount)
@@ -99,7 +93,7 @@ func TestEventStreamSource(t *testing.T) {
 			defer cancel()
 
 			spy := &EventTargetSpy{EventTarget: target}
-			src := &EventSource{spy}
+			src := &event.EventSource{spy}
 
 			events := src.Listen(ctx, "gost-event")
 
@@ -142,67 +136,4 @@ func (s *EventTargetSpy) RemoveEventListener(
 ) {
 	s.removeCallCount++
 	s.EventTarget.RemoveEventListener(t, h, opts...)
-}
-
-// EventSource embeds an [EventTarget] and provides events in a channel,
-// simplifying Go code consuming events.
-type EventSource struct {
-	event.EventTarget
-}
-
-type eventSourceOptions struct {
-	buf int
-}
-
-type EventSourceOption func(*eventSourceOptions)
-
-func BufSize(buf int) EventSourceOption {
-	return func(o *eventSourceOptions) { o.buf = buf }
-}
-
-// Listen adds an event listener for events of type t and returns a channel of
-// events containing all the events. Cancelling context ctx will remove the
-// event listener and close the channel. If no context is passed, the event
-// listener will never be removed.
-//
-// Ordering of events is guaranteed when the channel buffer is not full and all
-// events are dispatched from the same goroutine. The channel buffer size is
-// controlled with the [BufSize] option. Default value is [DEFAULT_BUF].
-func (s EventSource) Listen(
-	ctx context.Context,
-	t string,
-	opts ...EventSourceOption,
-) <-chan *event.Event {
-	opt := eventSourceOptions{buf: DEFAULT_BUF}
-	for _, o := range opts {
-		o(&opt)
-	}
-	c := make(chan *event.Event, opt.buf)
-	handler := event.NewEventHandlerFunc(func(e *event.Event) error {
-		// It is assumed that all events are dispatched from the same
-		// goroutine. If the buffer permits, send as a blocking call,
-		// ensuring ordering of events.
-		//
-		// If the channel is not ready to accept messages, send in a new
-		// goroutine to avoid blocking the EventTarget
-		select {
-		case c <- e:
-		default:
-			go func() { c <- e }()
-		}
-		return nil
-	})
-
-	s.EventTarget.AddEventListener(t, handler)
-
-	if ctx != nil {
-		// If no context is provided, the event listener is never removed.
-		go func() {
-			<-ctx.Done()
-			s.EventTarget.RemoveEventListener(t, handler)
-			close(c)
-		}()
-	}
-
-	return c
 }
